requirements.txt
click==8.1.6
contourpy==1.1.0
cycler==0.11.0
exceptiongroup==1.1.2
fonttools==4.42.0
gensim==4.3.1
iniconfig==2.0.0
joblib==1.3.2
kiwisolver==1.4.4
lxml==4.9.3
matplotlib==3.7.2
networkx==3.1
nltk==3.8.1
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
pyparsing==3.0.9
pytest==7.4.0
python-dateutil==2.8.2
python-docx==0.8.11
regex==2023.8.8
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smart-open==6.3.0
threadpoolctl==3.2.0
tomli==2.0.1
tqdm==4.66.1
wordcloud==1.9.2
